requirements.txt
crewai
crewai-tools
langchain
langchain-openai
synapseclient==4.8.0
setuptools
PyYAML
requests
tqdm 